Board에 관련된 Dto 위치 변경
