+++ src/main/java/study/board/board/BoardService.java
@@ -6,6 +6,9 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import study.board.board.dto.BoardCreateDto;
+import study.board.board.dto.BoardDto;
+import study.board.board.dto.BoardEditDto;
 import study.board.file.UploadFile;
 import study.board.file.UploadFileRepository;
 import study.board.file.UploadFilesRepository;
